pharmatechapis/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
daphne==4.2.0
dj-database-url==2.3.0
Django==5.1.6
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
Faker==33.1.0
firebase_admin==7.1.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
frozenlist==1.6.0
gcloud==0.18.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
kiwisolver==1.4.8
kombu==5.5.2
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
multidict==6.4.4
mysqlclient==2.2.7
numpy==2.2.3
oauth2_provider==0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.1
Pyrebase4==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jwt==4.1.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.0
zope.interface==7.2
